refactor: cut unnecessary declarations
After looking at TOP's solution, I realised that the links didn't have
to be bold so I cut that. I removed the explicit height declaration
(min-height) and adjusted some declaration values to match the
solution's.
I also decided to track the file '.vscode/settings.json'. I'm not
entirely sure what it does but I think it was generated when I changed my
color theme on VSCode so... got something to do with that, methinks.